Add the DJMax Studio shell and the Technika preview
DJMaxEditor.Studio (net8.0-windows) is a new WPF shell with a
retained-mode timeline, a TECHNIKA gameplay playfield sized from the
DMT2 client's own data, NAudio keysound playback with ring-out release,
and four headless probes for frame, audio, pause-tail and corpus
verification.

The WinForms editor gains a ptSequencer-style vertical timeline with
chart-driven overflow columns, Respect zone themes for 4B, 5B, 6B and
8B, a clickable time ruler, continuous follow-scroll, and uncapped
playback repaints.

Ships the application projects only: DJMaxEditor, DJMaxEditor.Core,
DJMaxEditor.Studio and UnpackMe.SDK.Standard. The test harnesses are
kept out of the solution and out of the repository.

+++ DJMaxEditor/Files/bytes/TQSaveFile.cs
@@ -88,6 +88,12 @@ namespace DJMaxEditor.Files.bytes
                     track.Write(BitConverter.GetBytes(eventsCount), 0, 4);
 
 
+                    // The (short) casts on the byte-wide fields below are load-bearing:
+                    // net8 added System.Half, and byte converts implicitly to both Half and
+                    // short, so an uncast GetBytes(evnt.Vel) is ambiguous when this shared
+                    // source is compiled into the studio shell. short is the overload .NET
+                    // Framework already resolved to, and only the low byte is written, so the
+                    // file bytes are unchanged.
                     foreach (EventData evnt in td.Events)
                     {
                         track.Write(BitConverter.GetBytes(evnt.Tick), 0, 4);
@@ -97,7 +103,7 @@ namespace DJMaxEditor.Files.bytes
                             case EventType.Volume:
                                 {
                                     track.Write(new byte[] { 2 }, 0, 1);
-                                    track.Write(BitConverter.GetBytes(evnt.Volume), 0, 1);
+                                    track.Write(BitConverter.GetBytes((short)evnt.Volume), 0, 1);
                                     track.Write(BitConverter.GetBytes(0), 0, 1);
                                     track.Write(BitConverter.GetBytes(0), 0, 1);
                                     track.Write(BitConverter.GetBytes(0), 0, 1);
@@ -116,9 +122,9 @@ namespace DJMaxEditor.Files.bytes
                                     }
 
                                     track.Write(BitConverter.GetBytes(insno), 0, 2);
-                                    track.Write(BitConverter.GetBytes(evnt.Vel), 0, 1);
-                                    track.Write(BitConverter.GetBytes(evnt.Pan), 0, 1);
-                                    track.Write(BitConverter.GetBytes(evnt.Attribute), 0, 1);
+                                    track.Write(BitConverter.GetBytes((short)evnt.Vel), 0, 1);
+                                    track.Write(BitConverter.GetBytes((short)evnt.Pan), 0, 1);
+                                    track.Write(BitConverter.GetBytes((short)evnt.Attribute), 0, 1);
                                     track.Write(BitConverter.GetBytes(evnt.Duration), 0, 2); // Is Duration 2 bytes? or 1 bytes
                                     track.Write(BitConverter.GetBytes(0), 0, 1);
